Configuring an endpoint for GetCourseDetailsRequest

+++ soap-course-management/src/main/java/net/jimmywin/courses/GetCourseDetailsRequest.java
@@ -47,13 +47,11 @@ public class GetCourseDetailsRequest {
 
     /**
      * Gets the value of the id property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link BigInteger }
-     *     
+     *
+     * @return possible object is
+     * {@link BigInteger }
      */
-    public BigInteger getId() {
+    public int getId() {
         return id;
     }
